break up interpreter for readability, add to docs
